♻️Move canvas loading to canvasStore
